# Preload the distance index to avoid very slow initial read mapping