@adamnovak 2023-02-09
The overall RPS numbers in my testing are still pretty bad, but I think a lot of that is that there aren't enough reads to give batches to each thread at the default batch size.